backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.93.0
anyio==4.13.0
asgiref==3.11.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
curl_cffi==0.15.0
distro==1.9.0
Django==6.0.4
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
frozendict==2.4.7
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
multitasking==0.0.12
numpy==2.4.4
packaging==26.0
pandas==3.0.2
peewee==4.0.4
platformdirs==4.9.6
prompt_toolkit==3.0.52
protobuf==7.34.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
redis==7.4.0
requests==2.33.1
rich==14.3.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
whitenoise==6.12.0
yfinance==1.2.1
